Agente Fiscal/requirements.txt
streamlit==1.29.0
crewai==0.28.8
crewai-tools==0.1.6
pandas==2.1.4
python-dotenv==1.0.0
pyyaml==6.0.1
langchain==0.1.10
langchain-community==0.0.25
requests==2.31.0
numpy==1.24.3